chore: remove version.txt
