+++ goreleaser_internal.yaml
@@ -12,7 +12,7 @@ builds:
       - amd64
       - arm64      
     ldflags:
-      - -s -w -X mgc/sdk/version.RawVersion={{.Version}}
+      - -s -w -X main.RawVersion=v0.1.23
     flags:
       - -tags=embed -buildvcs=true
     main: mgc/cli/main.go
